feature/refactor: add overlay, create burger-menu icon and menu-billet. Move mui-restyling variables into single file style-constants.ts
