+++ src/components/drop-down/drop-down.tsx
@@ -1,10 +1,24 @@
 import React, {FunctionComponent} from 'react';
-import {TableHeader} from '../table-header/table-header';
 
-export const DropDown: FunctionComponent = () => {
+import burgerStyles from './drop-down.module.css';
+
+import {Overlay} from '../overlay/overlay';
+
+import {TBurgerMenu} from '../../utils/types';
+
+
+export const DropDown: FunctionComponent<TBurgerMenu> = (props) => {
   return (
-    <div>
-
-    </div>
+    <>
+      {
+        props.menuIsOpen &&
+        <Overlay onClose={props.onClose}/>
+      }
+      <div className={props.menuIsOpen
+        ? `${burgerStyles['menu-body']} ${burgerStyles['menu-body_opened']}`
+        : `${burgerStyles['menu-body']}`}>
+        {props.children}
+      </div>
+    </>
   )
 }
