+++ src/components/header/header.module.css
@@ -1,7 +1,7 @@
 .header {
   width: 100%;
   height: 90px;
-  padding: 0 58px;
+  padding: 0 40px;
   color: #FFFFFF;
   background-color:#1A1B22;
   display: flex;
