+++ src/components/overlay/overlay.module.css
@@ -1,0 +1,17 @@
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100vw;
+  height: 100vh;
+  z-index: 3;
+  background: rgba(0, 0, 0, .3);
+}
+
+.body-overlay {
+  overflow: hidden;
+}
+
