+++ src/components/overlay/overlay.tsx
@@ -1,0 +1,9 @@
+import React, {FunctionComponent} from 'react';
+
+import overlayStyles from './overlay.module.css';
+
+export const Overlay: FunctionComponent<{onClose: () => void}> = (props) => {
+  return (
+    <div className={overlayStyles.overlay} onClick={props.onClose}></div>
+  )
+}
